Changes to HttpHandler and DefaultHttpHandler for flexibility
- Removed headers argument from HttpHandler.Callback#onResponse
- Moved Tls12SocketFactory into separate file to unclutter DefaultHttpHandler
- Deleted CachePolicy
- Added protected methods configureClient and configureRequest to DefaultHttpHandler for subclasses to override
- Updated demo MainActivity to use new DefaultHttpHandler methods
